// src/components/CardProfile/CardProfile.d.ts
export interface CardProps {
    classname?: string;
    title?: string;
    progress?: string;
    description?: string;
    courseimg?: string;
    buttonName?: string;
    textUnderBar?: string;
    currentClass?: number;
    globalCourse?: any;
    name?: string;
    profilecard?: boolean;
    lvl?: string;
    exp?: string;
    lvlprogress?: any;
}